chore: 타입을 types 폴더에 선언
